requirements.txt
asgiref==3.3.4
Django==3.2
gunicorn==20.1.0
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
